+++ apps/35mm-film-price-guide/components/HeaderStats.js
@@ -2,10 +2,10 @@ import React from "react";
 import styled from "styled-components";
 
 import { CURRENCY, DATE, FILM_PRICE_DATA, routes } from "../constants";
-import { c_black, c_grey_dark } from "../../../../constants/styles/colors";
-import { dateFromUnix } from "../../../../utils/time";
+import { c_black, c_grey_dark } from "../../../constants/styles/colors";
+import { dateFromUnix } from "../../../utils/time";
 import { filmPriceStats, generateAnchor } from "../utils";
-import Link from "../../../../core/components/controls/Link";
+import Link from "../../../core/components/controls/Link";
 
 const HeaderStats = styled.ul`
   display: ${props => props.hidden && "none"};
